+++ Mcsg.Story.Api/Services/StoryService.cs
@@ -92,7 +92,7 @@ public partial class StoryService : IStoryService
         return await _postService.Delete(postId);
     }
 
-    public async Task<ChapterResponse> GetChapter(string hashId, int order)
+    public async Task<ChapterResponse> GetChapter(string hashId, float order)
     {
         return await _postService.GetSeriesChapter(hashId, order);
     }
